Moved files, changed old usings #22
Moved TokenUserPairs.cs to /BLL
Moved Store.cs to /DAL
Deleted /Data folder
