+++ Gruvo/Gruvo/BLL/GruvoCookieHandler.cs
@@ -1,5 +1,3 @@
-
-﻿using Gruvo.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Filters;
